order-generator/requirements.txt
python-dateutil
confluent-kafka
azure-eventhub
azure-identity
websocket-client
dataclasses
